Profile

+++ Database_Hospital_Application/Models/Entities/Employee.cs
@@ -12,7 +12,7 @@ namespace Database_Hospital_Application.Models.Entities
         public int Id { get; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int BirthNumber { get; set; }
+        public long BirthNumber { get; set; }
         public SexEnum Sex { get; set; }
 
         public Department _department { get; set; }
